LayoutTests/platform/glib/imported/w3c/web-platform-tests/html/rendering/widgets/button-layout/flex-expected.txt


1
2
3
4

1
2
3
4

1
2
3
4

abc

abc

abc

PASS inline-flex
PASS flex
PASS align-items:flex-start should work
PASS align-items:stretch should work
FAIL No align-items should work as stretch assert_equals: expected 18 but got 48
